+++ crates/core/src/lib.rs
@@ -1,9 +1,15 @@
+pub mod effect;
 pub mod error;
 pub mod event;
 pub mod message;
+pub mod observer;
 pub mod tool_spec;
+pub mod traits;
 
+pub use effect::{CancelSpec, CrawlScope, CrawlTask, StatusSpec, ToolEffect, WaitSpec};
 pub use error::{RuntimeError, ToolError, ToolExecutionError};
 pub use event::AssistantEvent;
 pub use message::{ContentBlock, ConversationMessage, MessageRole, TokenUsage};
+pub use observer::RuntimeObserver;
 pub use tool_spec::ToolSpec;
+pub use traits::ToolExecutor;
